pyproject.toml
[tool.poetry]
name = "django-repo-structure"
version = "0.1.0"
description = ""
authors = ["Michael Asumadu <masumadu@archsolutions.online>"]
readme = "README.md"
packages = [{include = "django_repo_structure"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-environ = "^0.11.2"
pinject = "^0.14.1"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
coverage = "^7.5.4"
pre-commit = "^3.7.1"
cryptography = "^42.0.8"
jinja2 = "^3.1.4"
django-storages = "^1.14.4"
boto3 = "^1.34.144"
celery = "^5.4.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
